Refactor: #traverse method reduces repetition
Rubocop complained about too many lines in this
class. I went through the exercise of refactoring
so #traverse the helper method could be used to
reduce the code for traversing to the end of the
linked list.

It was a good experience to practise writing a
block and using yield. In the end, it turned out
to be unhelpful because I only ended up using it
for #contains? and #find since all the other
methods had simpler implementation using direct
traversal or they had edge cases that #traverse
didn't cleanly handle.